rename OmniMatchViewModel to OmniMatchUtil
