+++ app/src/main/java/com/dimagi/biometric/viewmodels/FaceMatchViewModel.java
@@ -5,6 +5,7 @@ import android.util.Log;
 
 import androidx.annotation.NonNull;
 
+import com.dimagi.biometric.OmniMatchUtil;
 import com.google.protobuf.InvalidProtocolBufferException;
 
 import java.io.IOException;
@@ -70,13 +71,13 @@ public class FaceMatchViewModel extends BaseTemplateViewModel {
             Log.e(TAG, "Failed to init face view model: " + ex.getResultCode());
         }
 
-        omniMatchViewModel = new OmniMatchViewModel();
+        omniMatchUtil = new OmniMatchUtil();
     }
 
     @Override
     public void insertRecord(MatcherCommon.Record record, String id) {
         try {
-            omniMatchViewModel.insertRecord(matcherNative, matcherInstance, record, id);
+            omniMatchUtil.insertRecord(matcherNative, matcherInstance, record, id);
         } catch (OmniMatchException ex) {
             Log.e(TAG, "Failed to insert finger record: " + ex.getResultCode());
         }
@@ -85,7 +86,7 @@ public class FaceMatchViewModel extends BaseTemplateViewModel {
     @Override
     public float verifyRecord(MatcherCommon.Record record, String id) {
         try {
-            Matcher.RecordsResult result = omniMatchViewModel.verifyRecord(matcherNative, matcherInstance, record, id);
+            Matcher.RecordsResult result = omniMatchUtil.verifyRecord(matcherNative, matcherInstance, record, id);
             return result.getResultsList().get(0).getCandidate().getScores().getFace().getScore();
         } catch (OmniMatchException ex) {
             Log.e(TAG, "Failed to verify face record: " + ex.getResultCode());
@@ -98,7 +99,7 @@ public class FaceMatchViewModel extends BaseTemplateViewModel {
     @Override
     public Matcher.RecordsResult identifyRecord(MatcherCommon.Record record, float threshold, int maxCandidates) {
         try {
-            return omniMatchViewModel.identifyRecord(matcherNative, matcherInstance, record, threshold, maxCandidates);
+            return omniMatchUtil.identifyRecord(matcherNative, matcherInstance, record, threshold, maxCandidates);
         } catch(OmniMatchException ex) {
             Log.e(TAG, "Failed to identify face record: " + ex.getResultCode());
         } catch (IOException ex) {
@@ -110,7 +111,7 @@ public class FaceMatchViewModel extends BaseTemplateViewModel {
     @Override
     public BioCommon.MatcherTemplate createTemplate(byte[] image, int position, Common.ImageFormat imageFormat) {
         try {
-            return omniMatchViewModel.createFaceTemplate(templateCreatorNNNative, templateCreatorNNInstance, image, imageFormat);
+            return omniMatchUtil.createFaceTemplate(templateCreatorNNNative, templateCreatorNNInstance, image, imageFormat);
         } catch (OmniMatchException | InvalidProtocolBufferException ex) {
             Log.e(TAG, "Error creating face template");
         }
@@ -120,7 +121,7 @@ public class FaceMatchViewModel extends BaseTemplateViewModel {
     @Override
     public MatcherCommon.Record createRecord(List<BioCommon.MatcherTemplate> templates) {
         if (templates.size() > 0) {
-            return omniMatchViewModel.createFaceRecord(templates.get(0));
+            return omniMatchUtil.createFaceRecord(templates.get(0));
         }
         return null;
     }
